+++ backend/src/SkillCraft.EntityFrameworkCore/Configurations/CasteConfiguration.cs
@@ -20,12 +20,9 @@ internal class CasteConfiguration : AggregateConfiguration<CasteEntity>, IEntity
     builder.HasIndex(x => x.Name);
     builder.HasIndex(x => x.Skill);
 
-    builder.Ignore(x => x.Traits);
-
     builder.Property(x => x.Name).HasMaxLength(Name.MaximumLength);
     builder.Property(x => x.Skill).HasMaxLength(byte.MaxValue).HasConversion(new EnumToStringConverter<Skill>());
     builder.Property(x => x.WealthRoll).HasMaxLength(Roll.MaximumLength);
-    builder.Property(x => x.TraitsSerialized).HasColumnName(SkillCraftDb.Castes.Traits.Name);
 
     builder.HasOne(x => x.World).WithMany(x => x.Castes)
       .HasPrincipalKey(x => x.WorldId).HasForeignKey(x => x.WorldId)
