+++ backend/src/SkillCraft.EntityFrameworkCore/Mapper.cs
@@ -77,17 +77,9 @@ internal class Mapper
       Id = source.Id,
       Description = source.Description,
       Skill = source.Skill,
-      WealthRoll = source.WealthRoll
+      WealthRoll = source.WealthRoll,
+      Features = source.GetFeatures()
     };
-
-    foreach (KeyValuePair<Guid, TraitEntity> trait in source.Traits)
-    {
-      destination.Traits.Add(new TraitModel(trait.Value.Name)
-      {
-        Id = trait.Key,
-        Description = trait.Value.Description
-      });
-    }
 
     MapAggregate(source, destination);
 
@@ -323,7 +315,7 @@ internal class Mapper
       Id = source.Id,
       Description = source.Description,
       Attributes = source.GetAttributes(),
-      Features = source.GetFeatures(),
+      Traits = source.GetTraits(),
       Languages = source.GetLanguages(ToLanguage, world),
       Names = source.GetNames(),
       Speeds = source.GetSpeeds(),
